Move saver & remover interfaces from storage utils bundle into component
