+++ src/Badger/Bundle/GameBundle/Unlocker/BadgeUnlocker.php
@@ -7,8 +7,8 @@ use Badger\Component\Game\Model\BadgeInterface;
 use Badger\Component\Game\Model\ClaimedBadgeInterface;
 use Badger\Component\Game\Factory\UnlockedBadgeFactoryInterface;
 use Badger\Component\Game\Repository\UnlockedBadgeRepositoryInterface;
-use Badger\StorageUtilsBundle\Remover\RemoverInterface;
-use Badger\StorageUtilsBundle\Saver\SaverInterface;
+use Badger\Component\StorageUtils\Remover\RemoverInterface;
+use Badger\Component\StorageUtils\Saver\SaverInterface;
 use Badger\Component\User\Model\UserInterface;
 
 /**
