+++ src/Badger/Bundle/StorageUtilsBundle/Doctrine/Saver/BaseSaver.php
@@ -4,7 +4,7 @@ namespace Badger\StorageUtilsBundle\Doctrine\Saver;
 
 use Doctrine\Common\Persistence\ObjectManager;
 use Doctrine\Common\Util\ClassUtils;
-use Badger\StorageUtilsBundle\Saver\SaverInterface;
+use Badger\Component\StorageUtils\Saver\SaverInterface;
 
 /**
  * Base saver, declared as different services for different classes
